refactor: add product folder for dto
